requirements.txt
absl-py==2.2.2
alembic==1.15.2
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2024.6.1
gast==0.6.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
Jinja2==3.1.4
joblib==1.5.0
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
MetaTrader5==5.0.4993
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
optuna==4.3.0
packaging==25.0
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sklearn-compat==0.1.3
SQLAlchemy==2.0.41
sympy==1.13.3
ta==0.11.0
ta-lib @ file:///C:/Users/arman/Desktop/EURUSD/ta_lib-0.6.3-cp310-cp310-win_amd64.whl#sha256=effb277b2470f44e41cd2ac350b2c73b790c6ca1f8f0cd65264f97358f1b8e24
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.14.1
pandas_ta   